Move a bunch of path logic into Core/Util/PathUtil.cpp/h
.
