+++ Core/Util/RecentFiles.cpp
@@ -9,6 +9,7 @@
 #include "Common/TimeUtil.h"
 #include "Core/Loaders.h"
 #include "Core/Util/RecentFiles.h"
+#include "Core/Util/PathUtil.h"
 #include "Core/Config.h"
 
 RecentFilesManager g_recentFiles;
